added: --mods
Nearly the same output as normal Viper, however slightly different
wording on the formatting error message, and no "Enabled mods" line, and
if a mod is missing a .Version value it's left blank.

It's too much hassle to go around and add edge cases for the missing
value, hence I won't. It is also significantly faster than normal Viper,
about double as fast, although most of the time spent by normal Viper is
just starting up, the actual code that returns the output is far faster
than the ViperSH equivalent.